SheshinProject/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
transformers==4.35.2
pillow==10.1.0
streamlit==1.28.2